채팅 코드 작성

+++ src/main/java/com/example/CoEduServer/dto/ChatDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ChatDTO {
     // 각 채널을 구분할 수 있는 id
-    private Integer channelId;
-    private Integer writerId;
+    private Long channelId;
+    private Long writerId;
     private String chat;
 }
